\name{logit}
\alias{logit}
\title{
  Computes the logit transform
}
\description{
  Computes the logit transform for objects of type numeric and 
  objects of class "pwr".   
}
\usage{
  logit(mu)
}
\arguments{
  \item{mu}{
    A real number on the interval [0, 1]
  }
}
\value{
  A numeric equal to the logit of mu, a real number.
}
\author{
  Grant Izmirlian <izmirlian at mail.nih.gov>
}
