\name{logitInv}
\alias{logitInv}
\title{
  Computes the inverse logit transform
}
\description{
  Computes the inverse logit transform for objects of type 
  numeric and objects of class "pwr".   
}
\usage{
  logitInv(eta)
}
\arguments{
  \item{eta}{
    Any real number 
  }
}
\value{
    A numeric equal to the logit inverse of mu, a real 
    number on the interval [0, 1]
}
\author{
  Grant Izmirlian <izmirlian at mail.nih.gov>
}
