\name{view.presentation}
\alias{view.presentation}
\title{
  Call this function (without arguments) to open a presentation which
  describes the theory behind the package and presents a simulation
  study.
}
\description{
  View a presentation
}
\usage{
  view.presentation()
}
\author{
  Grant Izmirlian <izmirlig at nih dot gov>
}
